requirements.txt
flask
httpx
truststore
httpx[http2]
logging
typing
fastapi
fastapi_mcp
uvicorn
mcp[cli]
mcp
scikit-learn
python-dotenv
PyJWT
python-jose